+++ bonelab/cli/apply_sitk_transform.py
@@ -9,7 +9,7 @@ from bonelab.util.time_stamp import message
 from bonelab.io.vtk_helpers import get_vtk_writer
 from bonelab.util.registration_util import create_file_extension_checker, create_string_argument_checker, INTERPOLATORS, \
     INPUT_EXTENSIONS, TRANSFORM_EXTENSIONS, check_inputs_exist, check_for_output_overwrite, read_image
-from bonelab.cli.demons_registration import IMAGE_EXTENSIONS
+from bonelab.util.demons_registration_util import IMAGE_EXTENSIONS
 
 
 def read_transform(fn: str, invert: bool, silent: bool) -> sitk.Transform:
